requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit
plotly
joblib
pyarrow
python-dotenv
requests